web_app/requirements.txt
Flask
python-dotenv
pandas
mysql-connector-python
openai
dotenv
requests